Bugfix and additions
The k nearest neighbours algorithm had a wrong branch condition for checking the KD Tree, making it run throu the whole tree

The UI is capable of zoom in and out with the scroll bar, up to a certain point. The right button of the mouce is used as
a reset for the zoom and the position of the UI, while the left is used for the Nearest Neighbour algorithm.

ToDo:
-> Implement the algorithm for cells.
-> COMMENTS!!!!!