Changed KD tree representation & more comments
KD tree was presented as a tree of pointers. Given its nature as a balanced binary tree, it
is now stored as an array that is traverersed as a binary tree. This format may allocate
additional unused space, but that does not create fragmentation to the memory.

Additional comments were added to the code to the existing and new code.