Adjust the algorithm for Cells and IOs & Changed the data structures
Drawbuffer input files:
=======================
A new format of files is parsed. This format is an extraction from
the ASP Tool, with the following form:
        <Name> <Type> <X-coord> <Y-coord> <Width> <Height>
the example files (drawbuffer_1.txt & drawbuffer_2.txt) have only
Cells, IOs, and the top module. In both cases the deign is flat.
More types of elements can be added in the future.

Data structures changes:
========================
The hash table that was used to store the point in replaced with an
1D array. This array is simillar to the Drawbuffer that is used in
ASP Tool. It is also used to store all the information extracted from
the Drawbuffer.

The sorting array is changed from an array of pointers to an array of
indexes to the array with the stored elements.

Changes to the Algorithm:
=========================
In order for the algorithm to handle these new elements, they have to
be represented in form more friendly to its current state. Therefore,
The Cells, IOs and Modules are represented to the algorithm with points
placed on their center. The coordinates of these points are not stored,
but they are calculated on the fly.